+++ packages/frontend/tests/helpers/pages/app.ts
@@ -1,6 +1,28 @@
 import { find, click, waitFor, fillIn, triggerKeyEvent } from '@ember/test-helpers';
+import {
+  create,
+  collection,
+  is,
+  count,
+  isVisible,
+  text,
+  fillable,
+  clickable,
+} from 'ember-cli-page-object';
 
 const toast = '[data-test-notification-message]';
+
+export const page = create({
+  hasNotificationPrompt: isVisible('[data-test-notification-prompt]'),
+  notificationPrompt: {
+    scope: '[data-test-notification-prompt]',
+    isVisible: isVisible(),
+    askLater: clickable('[data-test-ask-later]'),
+    askNever: clickable('[data-test-ask-never]'),
+    enable: clickable('[data-test-enable]'),
+    dismiss: clickable('[data-test-dismiss]'),
+  },
+});
 
 export const app = {
   selectors: {